requirements.txt
numpy
pillow
tensorflow==2.16.1
streamlit
requests